+++ Sofa/framework/Core/src/sofa/core/visual/DisplayFlags.cpp
@@ -26,237 +26,6 @@
 namespace sofa::core::visual
 {
 
-FlagTreeItem::FlagTreeItem(const std::string& showName, const std::string& hideName, FlagTreeItem* parent):
-    m_showName({showName}),
-    m_hideName({hideName}),
-    m_state(tristate::neutral_value),
-    m_parent(parent)
-{
-    if( m_parent ) m_parent->m_child.push_back(this);
-}
-
-
-void FlagTreeItem::addAliasShow(const std::string& newAlias)
-{
-    this->addAlias(this->m_showName, newAlias);
-}
-
-
-void FlagTreeItem::addAliasHide(const std::string& newAlias)
-{
-    this->addAlias(this->m_hideName, newAlias);
-}
-
-
-void FlagTreeItem::addAlias(sofa::type::vector<std::string>& name, const std::string& newAlias)
-{
-    name.push_back(newAlias);
-    return;
-}
-
-
-void FlagTreeItem::setValue(const tristate &state)
-{
-    this->m_state = state;
-    propagateStateDown(this);
-    propagateStateUp(this);
-}
-
-void FlagTreeItem::propagateStateDown(FlagTreeItem* origin)
-{
-    ChildIterator iter;
-    for( iter = origin->m_child.begin(); iter != origin->m_child.end(); ++iter)
-    {
-        (*iter)->m_state = origin->m_state;
-        propagateStateDown(*iter);
-    }
-}
-
-void FlagTreeItem::propagateStateUp(FlagTreeItem* origin)
-{
-    FlagTreeItem* parent = origin->m_parent;
-    if(!parent) return;
-
-    tristate flag = origin->m_state;
-    for( unsigned int i = 0 ; i < parent->m_child.size(); ++i)
-    {
-        FlagTreeItem* current = parent->m_child[i];
-        flag = fusion_tristate(current->m_state,flag);
-    }
-
-    parent->m_state=flag;
-    propagateStateUp(parent);
-}
-
-std::ostream& FlagTreeItem::write(std::ostream &os) const
-{
-    std::string s;
-    write_recursive(this,s);
-    s.erase(s.find_last_not_of(" \n\r\t")+1);
-    os << s;
-    return os;
-}
-
-void FlagTreeItem::showUnknownTokenMessage(const std::map<std::string, bool, FlagTreeItem::ci_comparison>& parseMap, std::string token) const
-{
-    sofa::type::vector<std::string> allFlagNames;
-    allFlagNames.reserve(parseMap.size());
-    for (const auto& [flagName, _] : parseMap)
-    {
-        allFlagNames.emplace_back(flagName);
-    }
-
-    std::stringstream tmp;
-    tmp << "Unknown flag '" << token << "'" << ". The closest existing ones:" << msgendl;
-    for(auto& [name, score] : sofa::helper::getClosestMatch(token, allFlagNames, 2, 0.6))
-    {
-        tmp << "\t" << "- " << name << " ("+ std::to_string((int)(100*score))+"% match)" << msgendl;
-    }
-    tmp << "Complete list is: " << allFlagNames;
-    msg_warning("DisplayFlags") << tmp.str() ;
-}
-
-std::istream& FlagTreeItem::read(std::istream &in)
-{
-    std::map<std::string, bool,  ci_comparison> parse_map;
-    create_parse_map(this,parse_map);
-    std::string token;
-    while(in >> token)
-    {
-        if( parse_map.find(token) != parse_map.end() )
-        {
-            std::map<std::string, bool>::const_iterator iter = parse_map.find(token);
-            std::string string1 = iter->first;
-
-            if(string1 != token)
-            {
-                msg_warning("DisplayFlags") << "Case of FlagTreeItem '" << token << "' is not correct, please use '"<< string1 <<"' instead";
-            }
-
-            parse_map[token] = true;
-        }
-        else
-        {
-            showUnknownTokenMessage(parse_map, token);
-        }
-    }
-    if( in.rdstate() & std::ios_base::eofbit ) { in.clear(); }
-
-    read_recursive(this,parse_map);
-    return in;
-}
-
-
-
-/*static*/
-void FlagTreeItem::create_parse_map(FlagTreeItem *root, std::map<std::string, bool, ci_comparison> &map)
-{
-    size_t sizeShow = root->m_showName.size();
-    size_t sizeHide = root->m_hideName.size();
-    for(size_t i=0; i<sizeShow; i++)
-    {
-        map[root->m_showName[i]] = false;
-    }
-    for(size_t i=0; i<sizeHide; ++i)
-    {
-        map[root->m_hideName[i]] = false;
-    }
-
-    ChildIterator iter;
-    for( iter = root->m_child.begin(); iter != root->m_child.end(); ++iter)
-    {
-        create_parse_map(*iter,map);
-    }
-}
-
-void FlagTreeItem::read_recursive(FlagTreeItem *root, const std::map<std::string, bool, ci_comparison> &parse_map)
-{
-    ChildIterator iter;
-    std::map<std::string,bool>::const_iterator iter_show;
-    std::map<std::string,bool>::const_iterator iter_hide;
-    for( iter = root->m_child.begin(); iter != root->m_child.end(); ++iter)
-    {
-        size_t sizeShow = (*iter)->m_showName.size();
-        size_t sizeHide = (*iter)->m_hideName.size();
-
-        bool found = false;
-
-        for(size_t i=0; i<sizeHide; i++)
-        {
-            iter_hide = parse_map.find((*iter)->m_hideName[i]);
-            if( iter_hide != parse_map.end() )
-            {
-                bool hide = iter_hide->second;
-                if(hide)
-                {
-                    if(i != 0)
-                    {
-                        msg_warning("DisplayFlags") << "FlagTreeItem '" << (*iter)->m_hideName[i] << "' is deprecated, please use '"<<(*iter)->m_hideName[0]<<"' instead";
-                    }
-
-                    (*iter)->setValue(tristate::false_value);
-                    found = true;
-                }
-            }
-        }
-        for(size_t i=0; i<sizeShow; i++)
-        {
-            iter_show = parse_map.find((*iter)->m_showName[i]);
-            if( iter_show != parse_map.end() )
-            {
-                bool show  = iter_show->second;
-                if(show)
-                {
-                    if(i != 0)
-                    {
-                        msg_warning("DisplayFlags") << "FlagTreeItem '" << (*iter)->m_showName[i] << "' is deprecated, please use '"<<(*iter)->m_showName[0]<<"' instead";
-                    }
-
-                    (*iter)->setValue(tristate::true_value);
-                    found = true;
-                }
-            }
-        }
-
-        if(!found)
-        {
-            (*iter)->m_state = tristate::neutral_value;
-            read_recursive(*iter,parse_map);
-        }
-    }
-}
-
-void FlagTreeItem::write_recursive(const FlagTreeItem* root, std::string& str )
-{
-    ChildConstIterator iter;
-    for( iter = root->m_child.begin(); iter != root->m_child.end(); ++iter )
-    {
-        switch( (*iter)->m_state.state )
-        {
-            case tristate::true_value:
-                str.append((*iter)->m_showName[0]);
-                str.append(" ");
-                break;
-            case tristate::false_value:
-                str.append((*iter)->m_hideName[0]);
-                str.append(" ");
-                break;
-            case tristate::neutral_value:
-                write_recursive(*iter,str);
-        }
-    }
-}
-
-std::ostream& operator<< ( std::ostream& os, const FlagTreeItem& root )
-{
-    return root.write(os);
-}
-std::istream& operator>> ( std::istream& in, FlagTreeItem& root )
-{
-    return root.read(in);
-}
-
-
 DisplayFlags::DisplayFlags():
     m_root(FlagTreeItem("showRoot","hideRoot",nullptr)),
     m_showAll(FlagTreeItem("showAll","hideAll",&m_root)),
