+++ Sofa/framework/Core/src/sofa/core/visual/DisplayFlags.h
@@ -23,129 +23,11 @@
 
 #include <sofa/core/config.h>
 #include <sofa/type/vector.h>
-#include <map>
 
+#include <sofa/core/visual/FlagTreeItem.h>
 
 namespace sofa::core::visual
 {
-
-struct SOFA_CORE_API tristate
-{
-    enum state_t { false_value, true_value, neutral_value } state;
-    tristate(bool b):state(b==true ? true_value : false_value )
-    {
-    }
-    tristate():state(true_value)
-    {
-    }
-    tristate(state_t state):state(state) {}
-
-    operator bool() const
-    {
-        return state == true_value ? true : false;
-    }
-
-    bool operator==(const tristate& t) const
-    {
-        return state == t.state;
-    }
-
-    bool operator!=(const tristate& t) const
-    {
-        return state != t.state;
-    }
-
-    bool operator==(const state_t& s) const
-    {
-        return state == s;
-    }
-
-    bool operator!=(const state_t& s) const
-    {
-        return state != s;
-    }
-
-    friend inline tristate fusion_tristate(const tristate& lhs, const tristate& rhs);
-    friend inline tristate merge_tristate(const tristate& previous, const tristate& current);
-    friend inline tristate difference_tristate(const tristate& previous, const tristate& current);
-};
-
-inline tristate fusion_tristate(const tristate &lhs, const tristate &rhs)
-{
-    if( lhs.state == rhs.state ) return lhs;
-    return tristate(tristate::neutral_value);
-}
-
-inline tristate merge_tristate(const tristate& previous, const tristate& current)
-{
-    if(current.state == tristate::neutral_value ) return previous;
-    return current;
-}
-inline tristate difference_tristate(const tristate& previous, const tristate& current)
-{
-    if( current.state == tristate::neutral_value || current.state == previous.state )
-        return tristate(tristate::neutral_value);
-    return current;
-}
-
-class SOFA_CORE_API FlagTreeItem
-{
-protected:
-    // Creating a case insensitive "find" function for map
-    struct ci_comparison
-    {
-        // case-independent (ci) comparison
-        struct nocase_compare
-        {
-            bool operator() (const unsigned char& c1, const unsigned char& c2) const
-            {
-                return tolower (c1) < tolower (c2);
-            }
-        };
-        bool operator() (const std::string & s1, const std::string & s2) const
-        {
-            return std::lexicographical_compare(s1.begin (), s1.end (), s2.begin (), s2.end (), nocase_compare ());
-        }
-    };
-
-    sofa::type::vector<std::string> m_showName;
-    sofa::type::vector<std::string> m_hideName;
-    tristate m_state;
-
-    FlagTreeItem* m_parent;
-    sofa::type::vector<FlagTreeItem*> m_child;
-
-    typedef type::vector<FlagTreeItem*>::iterator ChildIterator;
-    typedef type::vector<FlagTreeItem*>::const_iterator ChildConstIterator;
-
-public:
-    FlagTreeItem(const std::string& showName, const std::string& hideName, FlagTreeItem* parent = nullptr);
-
-    const tristate& state( ) const {return m_state;}
-    tristate& state() {return m_state;}
-
-    SOFA_CORE_API friend std::ostream& operator<< ( std::ostream& os, const FlagTreeItem& root );
-    SOFA_CORE_API friend std::istream& operator>> ( std::istream& in, FlagTreeItem& root );
-    std::ostream& write(std::ostream& os) const;
-    std::istream& read(std::istream& in);
-
-    void setValue(const tristate& state);
-
-    void addAliasShow(const std::string& newAlias);
-    void addAliasHide(const std::string& newAlias);
-    void addAlias(sofa::type::vector<std::string> &name, const std::string &newAlias);
-
-protected:
-    void propagateStateDown(FlagTreeItem* origin);
-    void propagateStateUp(FlagTreeItem* origin);
-    static std::map<std::string,bool, ci_comparison> create_flagmap(FlagTreeItem* root);
-    static void create_parse_map(FlagTreeItem* root, std::map<std::string,bool,ci_comparison>& map);
-    static void read_recursive(FlagTreeItem* root, const std::map<std::string,bool,ci_comparison>& map);
-    static void write_recursive(const FlagTreeItem* root,  std::string& str);
-
-    void showUnknownTokenMessage(const std::map<std::string, bool, FlagTreeItem::ci_comparison>& parseMap,
-                                 std::string token) const;
-};
 
 /** \brief Class which describes the display of components in a hierarchical fashion
 * DisplayFlags are conveyed by the VisualParams, and therefore are accessible in a
